BugFix in model

+++ Dev/DataModels/Atdi.DataModels.Sdrns.Server.Entities/Entities/IResLevels.cs
@@ -16,22 +16,22 @@ namespace Atdi.DataModels.Sdrns.Server.Entities
     [Entity]
     public interface IResLevels : IResLevels_PK
     {
-        double? ValueLvl { get; set; }
+        float? ValueLvl { get; set; }
         double? StddevLev { get; set; }
-        double? VMinLvl { get; set; }
-        double? VMMaxLvl { get; set; }
+        float? VMinLvl { get; set; }
+        float? VMMaxLvl { get; set; }
         double? LimitLvl { get; set; }
         double? OccupancyLvl { get; set; }
         double? PMinLvl { get; set; }
         double? PMaxLvl { get; set; }
         double? PDiffLvl { get; set; }
-        double? FreqMeas { get; set; }
-        double? ValueSpect { get; set; }
+        float? FreqMeas { get; set; }
+        float? ValueSpect { get; set; }
         double? StdDevSpect { get; set; }
         double? VMinSpect { get; set; }
         double? VMMaxSpect { get; set; }
         double? LimitSpect { get; set; }
-        double? OccupancySpect { get; set; }
+        float? OccupancySpect { get; set; }
         long? ResMeasId { get; set; }
         IResMeas RESMEAS { get; set; }
     }
